edge_agent/requirements.txt
paho-mqtt==1.6.1
tensorflow==2.15.0
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
